+++ activestorage/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Allow accepting `service` as a proc as well in `has_one_attached` and `has_many_attached`.
+
+    *Yogesh Khater*
 
 Please check [7-1-stable](https://github.com/rails/rails/blob/7-1-stable/activestorage/CHANGELOG.md) for previous changes.